Update deprecated function
